Update qops to include a update stack step for cookbooks and be more informative when it releases cookbooks.
