+++ lib/qops/version.rb
@@ -1,3 +1,3 @@
 module Qops
-  VERSION = '0.4.0'
+  VERSION = '0.5.0'
 end
